rename src to app
